End of Genre

+++ client/src/app/messages/messages.component.ts
@@ -23,20 +23,22 @@ export class MessagesComponent implements OnInit {
     this.loadMessages();
   }
 
-  loadMessages(){
-    this.loading = true;
-    this.messageService.getMessages(this.pageNumber,this.pageSize,this.container).subscribe(response =>{
+  loadMessages() {
+    // this.loading = true;
+    this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
       this.messages = response.result;
       this.pagination = response.pagination;
-      this.loading = false;
+      // this.loading = false;
     })
   }
 
   deleteMessage(id:number){
     this.confirmService.confirm('Confirm delete message','This cannot be undone').subscribe(result =>{
+      if(result) {
           this.messageService.deleteMessage(id).subscribe(()=>{
           this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
-      })
+        })
+      }   
     })
   }
